GUI: OSX puts DB/wallet in /Library/Application Support
Previously, the wallet and db were placed in the app itself.
Now put it in the OSX expected location.  Log files are also
moved out of the app itself.  A template config is packaged
with the app and created if it does not exist.

OSX users like to have the .app as one atomic unit, so
external configuration should not be required or created
for them.