DB: Store recent validator stats
Store the set of recent stats by block, so we can calculate
stats over a time range.

Currently this is defaulted to display last 7 days
in delegates panel.  But we can make this configurable

There's a lot of future work here how we present this
information.  Likely a new API method or parameter.

At 30 days, this should add ~2MB to the database
8 bytes * # blocks in 30 days * 3 stats possible